refactor: formalize & move `NonceDeduper` to p2p/utils pkg
