+++ p2p/utils/nonce_deduper_test.go
@@ -1,4 +1,4 @@
-package raintree
+package utils
 
 import (
 	"testing"
